maed it so that the response will vary in all service when a document is deleted based on if the id exists or not, before in certain services when a document was requested top be deleted, there wasnt any checks whether the id existed or not

+++ src/main/java/com/spring/SpeedAuction/security/Services/AuctionTypeCarService.java
@@ -61,8 +61,14 @@ public class AuctionTypeCarService {
     }
 
     public String deleteAuctionTypeCar(String id) {
-        auctionTypeCarRepository.deleteById(id);
-        return "AuctionTypeCar deleted";
+        AuctionTypeCar auctionTypeCar = auctionTypeCarRepository.findById(id).orElse(null);
+        if (auctionTypeCar != null) {
+            auctionTypeCarRepository.deleteById(id);
+            return "auctionTypeCar deleted";
+        }
+        else {
+            return "auctionTypeCar id doesnt exist";
+        }
     }
     public AuctionTypeCar createAuctiontypecar (CarDTO auctionTypeCarDTO) {
